requirements.txt
# API & web
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Data & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.2

# Scheduling & async
apscheduler==3.10.4
asyncio-throttle==1.0.2

# Alerting
slack-sdk==3.27.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0